% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb.R
\docType{data}
\name{availableEuPathDB}
\alias{availableEuPathDB}
\title{List of available curated species.}
\format{An object of class \code{NULL} of length 0.}
\usage{
availableEuPathDB
}
\description{
List of available curated species.
}
\keyword{datasets}
