% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb.R
\docType{data}
\name{build}
\alias{build}
\title{Boolean to see if packages should be built by devtools.}
\format{
An object of class \code{logical} of length 1.
}
\usage{
build
}
\description{
Boolean to see if packages should be built by devtools.
}
\keyword{datasets}
