% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb.R
\docType{data}
\name{build_dir}
\alias{build_dir}
\title{Set the build directory when creating packages and downloading data.}
\format{
An object of class \code{character} of length 1.
}
\usage{
build_dir
}
\description{
Set the build directory when creating packages and downloading data.
}
\keyword{datasets}
