% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_csv.R, R/check_s3.R
\name{check_csv}
\alias{check_csv}
\title{Check the metadata csv files and write only the 'good' entries.}
\usage{
check_csv(file_type = "OrgDb", bioc_version = "3.9",
  eu_version = "44")

check_csv(file_type = "OrgDb", bioc_version = "3.9",
  eu_version = "44")
}
\arguments{
\item{file_type}{Is this an OrgDB, GRanges, TxDb, OrganismDbi, or BSGenome dataset?}

\item{bioc_version}{Which bioconductor version is this for?}

\item{eu_version}{Which eupathdb version is this for?}

\item{file_type}{Is this an OrgDB, GRanges, TxDb, OrganismDbi, or BSGenome dataset?}

\item{bioc_version}{Which bioconductor version is this for?}

\item{eu_version}{Which eupathdb version is this for?}
}
\description{
While we are at it, put the failed entries into their own csv file so that I
can step through and look for why they failed.

While we are at it, put the failed entries into their own csv file so that I
can step through and look for why they failed.
}
