% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{check_files}
\alias{check_files}
\title{List the directory containing the various sqlite files and make sure they all have entries.}
\usage{
check_files(
  file_type = "OrgDb",
  bioc_version = NULL,
  eu_version = NULL,
  verbose = FALSE,
  destination = NULL
)
}
\arguments{
\item{file_type}{Is this an OrgDB, GRanges, TxDb, OrganismDbi, or BSGenome dataset?}

\item{bioc_version}{Which bioconductor version is this for?}

\item{eu_version}{Which eupathdb version is this for?}

\item{verbose}{Talk while running?}

\item{destination}{Place to put non-matched files.}
}
\description{
Any files which do not have csv entries should be deleted, but for the moment I will move them
to the current working directory in an attempt to learn about why they went wrong.
}
