% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_annotations.R
\name{check_post_result}
\alias{check_post_result}
\title{check that the result from a post makes sense.}
\usage{
check_post_result(result)
}
\description{
check that the result from a post makes sense.
}
