% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rda.R
\name{check_rda}
\alias{check_rda}
\title{Check the status of a saved database.
I save a bunch of rda files when downloading tables, this checks on them.}
\usage{
check_rda(name, entry, overwrite)
}
\arguments{
\item{name}{Table name}

\item{entry}{metadatum entry to query}

\item{overwrite}{when turned on, the existing file is deleted to make way for a new rda.}
}
\description{
Check the status of a saved database.
I save a bunch of rda files when downloading tables, this checks on them.
}
