% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_annotations.R
\name{choose_successful_post}
\alias{choose_successful_post}
\title{Choose a successful organism string when sending a POST}
\usage{
choose_successful_post(post_attempts, base_url)
}
\description{
Choose a successful organism string when sending a POST
}
