% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eupath_metadata.R
\name{download_eupath_metadata}
\alias{download_eupath_metadata}
\title{Returns metadata for all eupathdb organisms.}
\usage{
download_eupath_metadata(
  overwrite = FALSE,
  webservice = "eupathdb",
  bioc_version = NULL,
  dir = "EuPathDB",
  eu_version = NULL,
  write_csv = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{overwrite}{Overwrite existing data?}

\item{webservice}{Optional alternative webservice for hard-to-find species.}

\item{bioc_version}{Manually set the bioconductor release if desired.}

\item{dir}{Where to put the json.}

\item{eu_version}{Choose a specific eupathdb version?}

\item{write_csv}{Write a csv file in the format expected by AnnotationHubData?}

\item{verbose}{Print helper message about species matching?}
}
\value{
Dataframe with lots of rows for the various species in eupathdb.
}
\description{
Returns metadata for all eupathdb organisms.
}
\author{
Keith Hughitt
}
