% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{EupathDb}
\alias{EupathDb}
\alias{EupathDb-package}
\title{EupathDb: Some tools to create orgdb packages from the eupathdb.}
\description{
It should make them suitable for annotationhub, but I have not yet figured
out how Keith has linked them.
}
