% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{get_eupath_entry}
\alias{get_eupath_entry}
\title{Search the eupathdb metadata for a given species substring.}
\usage{
get_eupath_entry(species = "Leishmania major", webservice = "eupathdb")
}
\arguments{
\item{species}{String containing some reasonably unique text in the desired
species name.}

\item{webservice}{The EuPathDB webservice to query.}
}
\value{
A single row from the eupathdb metadata.
}
\description{
When querying the eupathdb, it can be difficult to hit the desired species.
This is confounded by the fact that there are very similar named species
across different EupathDB projects.  Thus function seeks to make it a bit
easier to find the actual dataset desired.  If the specific species is not
found, look for a reasonably approximation. stop() if nothing is found.
}
\author{
atb
}
