% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eupath_fields.R
\name{get_eupath_fields}
\alias{get_eupath_fields}
\title{Extract query-able fields from the EupathDb.}
\usage{
get_eupath_fields(webservice, excludes = NULL)
}
\arguments{
\item{webservice}{Eupathdb, tritrypdb, fungidb, etc...}

\item{excludes}{List of fields to ignore.}
}
\value{
List of parameters.
}
\description{
This parses the result of a query to Eupath's webservice:
'GenesByMolecularWeight' and uses it to get a list of fields which are
acquireable elsewhere.
}
