% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_annotations.R
\name{get_eupath_gene_types}
\alias{get_eupath_gene_types}
\title{Attempt to get a list of sequence types.}
\usage{
get_eupath_gene_types(webservice = NULL)
}
\arguments{
\item{webservice}{choose a service to download from.}
}
\description{
Attempt to get a list of sequence types.
}
