% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eupath_pkgnames.R
\name{get_eupath_pkgnames}
\alias{get_eupath_pkgnames}
\title{Generate standardized package names for the various eupathdb species.}
\usage{
get_eupath_pkgnames(entry, version = NULL, column = "TaxonUnmodified")
}
\arguments{
\item{entry}{A metadatum entry.}

\item{version}{Choose a specific version of the eupathdb, only really useful
when downloading files.}
}
\value{
List of package names and some booleans to see if they have already
  been installed.
}
\description{
This is a surprisingly difficult problem.  Many species names in the eupathdb
have odd characters in the species suffix which defines the strain ID.  Many
of these peculiarities result in packages which are non-viable for
installation.  Thus this function attempts to filter them out and result in
consistent, valid package names.  They are not exactly the same in format as
other orgdb/txdb/etc packages, as I include in them a field for the eupathdb
version used; but otherwise they should be familiar to any user of the sqlite
based organism packages.
}
\details{
The default argument for this function shows the funniest one I have found so
far thanks to the hash character in the strain definition.
}
\author{
atb
}
