% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_filename.R
\name{get_metadata_filename}
\alias{get_metadata_filename}
\title{Create a list of output csv filenames containing the metadata for each datatype.}
\usage{
get_metadata_filename(
  webservice,
  bioc_version,
  eu_version,
  file_type = "valid",
  build_dir = "build"
)
}
\description{
Create a list of output csv filenames containing the metadata for each datatype.
}
