% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservices.R
\name{get_orthologs_all_genes}
\alias{get_orthologs_all_genes}
\title{Query ortholog tables from the eupathdb one gene at a time.}
\usage{
get_orthologs_all_genes(species = "Leishmania major", dir = "eupathdb",
  gene_ids = NULL, entry = NULL, ...)
}
\arguments{
\item{species}{What species to query?}

\item{dir}{Directory to which to save intermediate data (currently unused).}

\item{gene_ids}{List of gene IDs to query.}

\item{entry}{An entry from the eupathdb metadata to use for other parameters.}

\item{...}{Extra parameters for downloading eupathdb metadata.}
}
\description{
Querying the full ortholog table at eupathdb.org fails mysteriously.
This is a horrible brute-force approach to get around this.
}
