% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orthologs_one_gene.R
\name{get_orthologs_one_gene}
\alias{get_orthologs_one_gene}
\title{This peculiar and slow querying of orthologs is due to me crashing the
eupathdb web servers.}
\usage{
get_orthologs_one_gene(entry = NULL, gene = "LmjF.01.0010",
  dir = "EuPathDB")
}
\arguments{
\item{entry}{Metadata entry.}

\item{gene}{What gene to query?}

\item{dir}{Where to put the checkpoint file?}
}
\value{
table of orthologs for our one gene.
}
\description{
Therefore, I wrote this, which queries one gene at a time.  I think it would
be nice to change this to query multiple genes at a time.
}
