% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rda.R
\name{get_rda_filename}
\alias{get_rda_filename}
\title{Provide the filename for a temporary rda file for saving data while running.}
\usage{
get_rda_filename(entry, name)
}
\description{
Provide the filename for a temporary rda file for saving data while running.
}
