% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_semantic_columns.R
\name{get_semantic_columns}
\alias{get_semantic_columns}
\title{Haha this is nasty, extract interesting data columns from the eupathdb wall of json.}
\usage{
get_semantic_columns(webservice = "tritrypdb", regexes = NULL)
}
\arguments{
\item{webservice}{Where to download from.}

\item{regexes}{A set of regexes used to prune uninteresting columns.}
}
\description{
Haha this is nasty, extract interesting data columns from the eupathdb wall of json.
}
