% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eupath_metadata.R
\name{get_tags}
\alias{get_tags}
\title{Set up the taxonomy Tags for the various EuPathDB sub-projects.}
\usage{
get_tags()
}
\description{
Set up the taxonomy Tags for the various EuPathDB sub-projects.
}
