% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_versions.R
\name{get_versions}
\alias{get_versions}
\title{Download the current eupathdb version.}
\usage{
get_versions(eu_version = NULL, bioc_version = NULL)
}
\description{
My favorite site is of course tritrypdb, so get it there.
}
