% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ah_annotations.R
\name{load_ah_annotations}
\alias{load_ah_annotations}
\title{Shortcut for loading annotation data from AnnotationHub, making some EupathDB assumptions.}
\usage{
load_ah_annotations(species = "Leishmania major strain Friedlin",
  service = "TriTrypDB", type = "OrgDb", eu_version = NULL,
  wanted_fields = NULL)
}
\arguments{
\item{species}{String containing a unique portion of the desired species.}

\item{service}{Which eupath webservice is desired?}

\item{type}{Data type to load.}

\item{eu_version}{Gather data from a specific eupathdb version?}

\item{wanted_fields}{If not provided, this will gather all columns starting
with 'annot'.}
}
\value{
Big huge data frame of annotation data.
}
\description{
Shortcut for loading annotation data from AnnotationHub, making some EupathDB assumptions.
}
