% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eupath_annotations.R
\name{load_eupath_annotations}
\alias{load_eupath_annotations}
\title{Shortcut for loading annotation data from a eupathdb-based orgdb.}
\usage{
load_eupath_annotations(species = "Leishmania major",
  webservice = "tritrypdb", eu_version = NULL, wanted_fields = NULL)
}
\arguments{
\item{species}{String containing a unique portion of the desired species.}

\item{webservice}{Which eupath webservice is desired?}

\item{eu_version}{Gather data from a specific eupathdb version?}

\item{wanted_fields}{If not provided, this will gather all columns starting
with 'annot'.}
}
\value{
Big huge data frame of annotation data.
}
\description{
Every time I go to load the annotation data from an orgdb for a parasite, it
takes me an annoyingly long time to get the darn flags right.  As a result I
wrote this to shortcut that process.  Ideally, one should only need to pass
it a species name and get out a nice big table of annotation data.
}
