% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{load_eupath_go,list-method}
\alias{load_eupath_go,list-method}
\title{Load orgdb GO data if a pkglist is provided.}
\usage{
\S4method{load_eupath_go}{list}(
  query,
  webservice = "tritrypdb",
  eu_version = NULL,
  wanted_fields = NULL,
  gene_ids = NULL,
  columns = c("go", "evidence")
)
}
\arguments{
\item{x}{The list result of make_eupath_orgdb.}
}
\description{
Load orgdb GO data if a pkglist is provided.
}
