% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eupath_orthologs.R, R/load_eupath_pkg.R
\name{load_eupath_pkg}
\alias{load_eupath_pkg}
\title{Loads a pkg into the current R environment.}
\usage{
load_eupath_pkg(name, webservice = "eupathdb")

load_eupath_pkg(name, webservice = "eupathdb")
}
\arguments{
\item{name}{Package name.}

\item{webservice}{From where to get the package name.}
}
\description{
Loads a pkg into the current R environment.

Loads a pkg into the current R environment.
}
