% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_methods.R
\name{load_orgdb_annotations,list-method}
\alias{load_orgdb_annotations,list-method}
\title{Load orgdb data if a pkglist is provided.}
\usage{
\S4method{load_orgdb_annotations}{list}(
  orgdb = NULL,
  gene_ids = NULL,
  include_go = FALSE,
  keytype = "gid",
  location_column = "annot_location_text",
  type_column = "annot_gene_type",
  name_column = "annot_gene_product",
  fields = NULL,
  sum_exon_widths = FALSE
)
}
\arguments{
\item{x}{The list result of make_eupath_orgdb.}
}
\description{
Load orgdb data if a pkglist is provided.
}
