% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_granges.R
\name{make_eupath_granges}
\alias{make_eupath_granges}
\title{Generate a GRanges rda savefile from a gff file.}
\usage{
make_eupath_granges(entry = NULL, dir = "EuPathDB", version = NULL,
  copy_s3 = FALSE)
}
\arguments{
\item{entry}{Metadatum entry.}

\item{dir}{Place to put the resulting file(s).}

\item{version}{Optionally request a specific version of the gff file.}
}
\description{
There is not too much else to say. This uses import.gff from rtracklayer.
I should probably steal my code from hpgltools to make this work for any
version of a gff file, but the eupathdb is good about keeping consistent on
this front.
}
