% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_organismdbi.R
\name{make_eupath_organismdbi}
\alias{make_eupath_organismdbi}
\title{Create an organismDbi instance for an eupathdb organism.}
\usage{
make_eupath_organismdbi(
  entry = NULL,
  eu_version = NULL,
  installp = TRUE,
  reinstall = FALSE,
  exclude_join = "ENTREZID",
  copy_s3 = FALSE,
  build = TRUE
)
}
\arguments{
\item{entry}{A row from the eupathdb metadataframe.}

\item{eu_version}{Which version of the eupathdb to use for creating this package?}

\item{installp}{Install the resulting package?}

\item{reinstall}{Overwrite existing data files?}

\item{exclude_join}{I had a harebrained idea to automatically set up the
joins between columns of GO.db/reactome.db/orgdb/txdb objects.  This
variable is intended to exclude columns with common IDs that might
multi-match spuriously -- I think in the end I killed the idea though,
perhaps this should be removed or resurrected.}

\item{copy_s3}{Copy the 2bit file into an s3 staging directory for copying to AnnotationHub?}
}
\value{
The result of attempting to install the organismDbi package.
}
\description{
The primary goal of an organismdbi instance is to provide a series of links
between an orgdb, txdb, and other relevant annotation packages
(reactome/go/etc).  In its current iteration, this function brings together a
couple columns from the orgdb, txdb, GO.db, and reactome.db.
}
\author{
Keith Hughitt, modified by atb.
}
