% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_orgdb.R
\name{make_eupath_orgdb}
\alias{make_eupath_orgdb}
\title{Create an orgdb SQLite database from the tables in eupathdb.}
\usage{
make_eupath_orgdb(entry = NULL, workdir = "EuPathDB",
  installp = TRUE, kegg_abbreviation = NULL, reinstall = FALSE,
  overwrite = FALSE, copy_s3 = FALSE, do_go = TRUE,
  do_goslim = TRUE, godb_source = NULL, do_orthologs = TRUE,
  do_interpro = TRUE, do_linkout = TRUE, do_pubmed = TRUE,
  do_pathway = TRUE, do_kegg = TRUE, do_uniprot = FALSE)
}
\arguments{
\item{entry}{If not provided, then species will get this, it contains all the information.}

\item{workdir}{Where to put all the various temporary files.}

\item{installp}{Install the resulting package?}

\item{kegg_abbreviation}{If known, provide the kegg abbreviation.}

\item{reinstall}{Re-install an already existing orgdb?}

\item{overwrite}{Overwrite a partial installation?}

\item{copy_s3}{Copy the 2bit file into an s3 staging directory for copying to AnnotationHub?}

\item{do_go}{Create the gene ontology table?}

\item{do_goslim}{Create the GOSLIM gene ontology table?}

\item{godb_source}{Which table to use for the putative union of the GO tables.}

\item{do_orthologs}{Create the gene ortholog table?}

\item{do_interpro}{Create the interpro table?}

\item{do_linkout}{Create a table of linkout data?}

\item{do_pubmed}{Create a table of pubmed entries?}

\item{do_pathway}{Create the pathway table?}

\item{do_kegg}{Attempt to create the kegg table?}

\item{do_uniprot}{Attempt to create a uniprot table?}
}
\value{
Currently only the name of the installed package.  This should
  probably change.
}
\description{
This function has passed through multiple iterations as the preferred
method(s) for accessing data in the eupathdb has changed.  It currently uses
my empirically defined set of queries against the eupathdb webservices.  As a
result, I have made some admittedly bizarre choices when creating the
queries.  Check through eupath_webservices.r for some amusing examples of how
I have gotten around the idiosyncrasies in the eupathdb.  Final note, I confirmed
with Cristina that it is not possible to acquire data specific to a given version
of the eupathdb.
}
\author{
Keith Hughitt with significant modifications by atb.
}
