% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_txdb.R
\name{make_eupath_txdb}
\alias{make_eupath_txdb}
\title{Generate an EuPathDB organism TxDb package.}
\usage{
make_eupath_txdb(entry = NULL, workdir = "EuPathDB",
  eu_version = NULL, reinstall = FALSE, installp = TRUE,
  copy_s3 = FALSE)
}
\arguments{
\item{entry}{One row from the organism metadata.}

\item{workdir}{Base directory for building the package.}

\item{eu_version}{Which version of the eupathdb to use for creating this package?}

\item{reinstall}{Overwrite an existing installed package?}

\item{installp}{Install the resulting package?}

\item{copy_s3}{Copy the 2bit file into an s3 staging directory for copying to AnnotationHub?}
}
\value{
TxDb instance name.
}
\description{
This will hopefully create a txdb package and granges savefile for a single
species in the eupathdb.  This depends pretty much entirely on the successful
download of a gff file from the eupathdb.
}
\author{
Keith Hughitt with significant modifications by atb.
}
