% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_annotations.R
\name{make_get_query}
\alias{make_get_query}
\title{Create a GET query to try out.
Here is the string theoretically expected for a GET:
Note that it is also problematic because of its inconsistent use of html encoded quotes and brackets.
https://fungidb.org/fungidb/service/record-types/transcript/searches/GenesByTaxon/reports/standard?organism=%5B%22Allomyces%20macrogynus%20ATCC%2038327%22%2C%22Blastocladiomycota%22%5D&reportConfig={"attributes":["primary_key"],"tables":[]}
   get_string <- glue(
      '{base_url}?organism=%5B%22{species_coded}%22%5D&reportConfig={{"attributes":[{group_string}],"tables":[]}')
   result <- httr::GET(url = get_string)}
\usage{
make_get_query()
}
\description{
Create a GET query to try out.
Here is the string theoretically expected for a GET:
Note that it is also problematic because of its inconsistent use of html encoded quotes and brackets.
https://fungidb.org/fungidb/service/record-types/transcript/searches/GenesByTaxon/reports/standard?organism=%5B%22Allomyces%20macrogynus%20ATCC%2038327%22%2C%22Blastocladiomycota%22%5D&reportConfig={"attributes":["primary_key"],"tables":[]}
   get_string <- glue(
      '{base_url}?organism=%5B%22{species_coded}%22%5D&reportConfig={{"attributes":[{group_string}],"tables":[]}')
   result <- httr::GET(url = get_string)
}
