% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_taxon_names.R
\name{make_taxon_names}
\alias{make_taxon_names}
\title{Iterate through the various ways of representing taxon names}
\usage{
make_taxon_names(entry, column = "TaxonUnmodified", spaces = "dot")
}
\arguments{
\item{entry}{An entry of the eupathdb metadata.}

\item{column}{Which column should be used to query the species name?}
}
\value{
A list of hopefully valid nomenclature names to be used elsewhere in
 this family.
}
\description{
Spend some time making sure they are valid, too.  Thus we want to get rid of
weird characters like hash marks, pipes, etc.
}
\author{
atb
}
