% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_final_package.R
\name{move_final_bsgenome_package}
\alias{move_final_bsgenome_package}
\title{FIXME: Not completed yet.  This should move the intermediates and
final bsgenomes.}
\usage{
move_final_bsgenome_package(pkglist, build_dir = "build")
}
\arguments{
\item{pkglist}{Package name of the bsgenome.}

\item{build_dir}{dirname of the build prefix.}
}
\description{
When completed, this will make certain that all the BSgenome pieces
will end up in the correct locations.
}
