% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_final_package.R
\name{move_final_granges_package}
\alias{move_final_granges_package}
\title{FIXME: Make sure this is correct. It should move granges
information to a final location.}
\usage{
move_final_granges_package(pkglist, build_dir = "build")
}
\arguments{
\item{pkglist}{Package name.}

\item{build_dir}{dirname of the build tree.}
}
\description{
FIXME: Make sure this is correct. It should move granges
information to a final location.
}
