% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_final_package.R
\name{move_final_organismdbi_package}
\alias{move_final_organismdbi_package}
\title{FIXME: Check that this is correct!}
\usage{
move_final_organismdbi_package(pkglist, build_dir = build_dir)
}
\arguments{
\item{pkglist}{Package name.}

\item{build_dir}{dirname of the build tree.}
}
\description{
This should move the organismDBI data into its final locations.
}
