% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_final_package.R
\name{move_final_orgdb_package}
\alias{move_final_orgdb_package}
\title{Move a package file to its final location for collation by
AnnotationHubData.}
\usage{
move_final_orgdb_package(pkglist, build_dir = "build")
}
\arguments{
\item{pkglist}{Information list from a make_pkg function.}

\item{build_dir}{base working directory.}

\item{type}{Which type of package is this?}
}
\description{
I want to have one location for the tarballs generated by
devtools::build() and another for the sqlite files which would be
used by AnnotationHub.  It can be a little tricky to get these
correct because of the various ways in which the built packages
might fail or need to be restarted.
}
