% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_final_package.R
\name{move_final_txdb_package}
\alias{move_final_txdb_package}
\title{FIXME: Complete this.  It should move the final TxDB data.}
\usage{
move_final_txdb_package(pkglist, build_dir = "build")
}
\arguments{
\item{pkglist}{Package name.}

\item{build_dir}{dirname of the build tree.}
}
\description{
FIXME: Complete this.  It should move the final TxDB data.
}
