% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_webservices.R
\name{post_eupath_annotations}
\alias{post_eupath_annotations}
\title{Gather all available annotation data for a given eupathdb species.}
\usage{
post_eupath_annotations(entry = NULL, dir = "EuPathDB",
  overwrite = FALSE)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{dir}{A directory into which to write the intermediate savefile.}

\item{overwrite}{If a partial table exists, overwrite it?}
}
\value{
A big honking table.
}
\description{
This function fills in the parameters to post_eupath_raw() so that one can
download all the available data for a given parasite into one massive table.
It should also provide some constraints to the data rather than leaving it
all as characters.  Caveat:  I manually filled in the list 'field_list' to
include the variable names and their text associations.  This is likely to
change in future releases of the tritrypdb.  It is probably possible to
automagically fill it in.  In addition, I am using GenesByMolecularWeight to
get the data, which is a bit weird.
}
