% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservices.R
\name{post_eupath_go_table}
\alias{post_eupath_go_table}
\title{Use the post interface to get GO data.}
\usage{
post_eupath_go_table(entry = NULL, dir = "eupathdb")
}
\arguments{
\item{entry}{The full annotation entry.}

\item{dir}{FIXME: I want to write some intermediate data to dir in case of
transient error.}
}
\value{
A big honking table.
}
\description{
Use the post interface to get GO data.
}
