% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_goslim.R
\name{post_eupath_goslim_table}
\alias{post_eupath_goslim_table}
\title{Use the POST interface to get GO data from the EuPathDB.}
\usage{
post_eupath_goslim_table(
  entry,
  working_species,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{overwrite}{Overwrite intermediate savefiles in case of incomplete install?}

\item{build_dir}{Location to write savefiles.}
}
\value{
A big honking table.
}
\description{
Use the POST interface to get GO data from the EuPathDB.
}
