% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_linkout.R
\name{post_eupath_linkout_table}
\alias{post_eupath_linkout_table}
\title{Use the post interface to get linkout data.}
\usage{
post_eupath_linkout_table(
  entry,
  working_species,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{overwrite}{Overwrite the savefile when attempting a redo?}

\item{build_dir}{Location to which to save intermediate savefile.}
}
\value{
A big honking table.
}
\description{
Use the post interface to get linkout data.
}
