% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_webservices.R
\name{post_eupath_ortholog_table}
\alias{post_eupath_ortholog_table}
\title{Use the post interface to get ortholog data.}
\usage{
post_eupath_ortholog_table(entry = NULL, dir = "EuPathDB",
  table = "OrthologsLite", overwrite = FALSE)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{dir}{Location to which to save an intermediate savefile.}

\item{table}{This defaults to the 'OrthologsLite' table, but that does not
exist at all eupathdb subprojects.}

\item{overwrite}{Overwrite incomplete savefiles?}
}
\value{
A big honking table.
}
\description{
Unfortunately, this function more often then not leads to a crash of the
eupathdb webservers.  As a result, I wrote a GET version of this which
iterates one gene at a time.
}
