% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_pdb.R
\name{post_eupath_pdb_table}
\alias{post_eupath_pdb_table}
\title{Use the post interface to get linkout data.}
\usage{
post_eupath_pdb_table(entry = NULL, build_dir = "EuPathDB", overwrite = FALSE)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{build_dir}{Location to which to save intermediate savefile.}

\item{overwrite}{Overwrite the savefile when attempting a redo?}
}
\value{
A big honking table.
}
\description{
Use the post interface to get linkout data.
}
