% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_table.R
\name{post_eupath_table}
\alias{post_eupath_table}
\title{Queries one of the EuPathDB APIs using a POST request.}
\usage{
post_eupath_table(entry, tables = "GOTerms", table_name = NULL, minutes = 30)
}
\arguments{
\item{entry}{The single metadatum containing the base url of the provider, species, etc.}

\item{table_name}{The name of the table to extract, this is provided to make
for prettier labeling.}

\item{minutes}{A timeout when querying the eupathdb.}

\item{query_body}{String of additional query arguments}
}
\value{
list containing response from API request.

More information
----------------
1. https://tritrypdb.org/tritrypdb/serviceList.jsp
}
\description{
This should return a dataframe representation of one table at the eupathdb.
It should also simplify the column names into something a bit more
consistent.
}
\author{
Keith Hughitt
}
