% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_map.R
\name{prefix_map}
\alias{prefix_map}
\title{A few webservices at the eupathdb are not what one would expect.}
\usage{
prefix_map(prefix)
}
\arguments{
\item{prefix}{Webservice to query.}
}
\description{
This maps the service name to the correct hostname of the webserver.
}
