% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.downloaded_metadata}
\alias{print.downloaded_metadata}
\title{Print a representation of some downloaded metadata.}
\usage{
\method{print}{downloaded_metadata}(x)
}
\arguments{
\item{x}{The metadata!}
}
\description{
Print a representation of some downloaded metadata.
}
