% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_print.R
\name{print.eupath_orgdb}
\alias{print.eupath_orgdb}
\title{Print a representation of a created eupath orgdb instance.}
\usage{
\method{print}{eupath_orgdb}(x)
}
\arguments{
\item{x}{Orgdb dataset}
}
\description{
Print a representation of a created eupath orgdb instance.
}
