% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_s3_ah.R
\name{query_s3_ah}
\alias{query_s3_ah}
\title{As yet another test, this function will download all the AH data one species at a time.}
\usage{
query_s3_ah(
  testing = TRUE,
  file_type = "OrgDb",
  cachedir = "~/scratch/eupathdb/cache"
)
}
\arguments{
\item{testing}{Use the annotationHub TESTING service rather than production.}

\item{file_type}{Type of data to query.}

\item{cachedir}{Place to put the downloaded files, useful for if one's homedirectory is too small.}
}
\description{
This may be run after the data has been uploaded to the annotationhub to
ensure that all the uploaded files are actually functional.
}
