% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_orgdb.R
\name{remove_eupath_nas}
\alias{remove_eupath_nas}
\title{Get rid of spurious NA entries in a table from the eupathdb.}
\usage{
remove_eupath_nas(table, name = "annot")
}
\description{
If these are not removed, creating the sqlite will fail.
}
