% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb.R
\name{start_eupathdb}
\alias{start_eupathdb}
\alias{availableEuPathDB}
\title{Get started with EuPathDB}
\usage{
start_eupathdb()
}
\value{
Used for its side-effect of opening the package vignette. A
        vector of experiment identifiers.
}
\description{
Get started with EuPathDB
}
\examples{
start_eupathdb()
}
\author{
Keith Hughitt
}
