% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xref_species.R
\name{xref_gidb_species}
\alias{xref_gidb_species}
\title{Cross reference information from the taxonDB vs. the downloaded metadata.}
\usage{
xref_gidb_species(
  metadatum,
  all_taxa_ids,
  taxon_number_column = "TaxonomyID",
  verbose = verbose
)
}
\arguments{
\item{metadatum}{The downloaded metadata for an individual species.}

\item{all_taxa_ids}{The taxonDB result.}

\item{taxon_number_column}{The metadatum column which should contain the relevant number.}

\item{verbose}{Be a chatty catty?}
}
\description{
Cross reference information from the taxonDB vs. the downloaded metadata.
}
