% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xref_species.R
\name{xref_species}
\alias{xref_species}
\title{Cross reference the taxonomy data from AnnotationHubData::getSpeciesList()}
\usage{
xref_species(valid, invalid, verbose = FALSE)
}
\arguments{
\item{valid}{Dataframe of entries which have thus far been deemed 'valid' by my tests.}

\item{invalid}{Dataframe of entries which failed.}

\item{verbose}{Print some information about what is found?}
}
\value{
Likely smaller data frame of valid information and larger dataframe of invalid.
}
\description{
Previously, the logic of this function resided in download_eupath_metadata(),
but I want to be able to test and poke at it separately to more effectively
ensure as many taxa as possible pass.  Therefore, I split it into its own
function. The secondary function of this is to set the 'Species' column as
appropriately as possible.
}
