% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xref_taxonomy.R
\name{xref_taxonomy_number}
\alias{xref_taxonomy_number}
\title{Cross reference the taxonomy data from GenomeInfoDb with EuPathDB metadata.}
\usage{
xref_taxonomy_number(
  metadatum,
  all_taxa_ids,
  verbose = FALSE,
  taxon_number_column = "TaxonomyID",
  metadata_taxon_column = "TaxonUnmodified"
)
}
\arguments{
\item{verbose}{Print some information about what is found as this runs?}

\item{metadata}{Information provided by downloading the metadata from a eupathdb sub project.}

\item{species_column}{Because the species column name has changed.}

\item{taxon_column}{Because the taxonomy column name has changed.}
}
\value{
List containing entries which pass and fail after xrefing against loadTaxonomyDb().
}
\description{
Previously this tried to set species and taxonomy number, but I am
having it only do the number now in the hopes that it simplifies
things somewhat.
}
