% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_uniprot_annotations.R
\name{download_uniprot_annotations}
\alias{download_uniprot_annotations}
\title{Make a table of uniprot anntotations when possible.}
\usage{
download_uniprot_annotations(gids, entry, dir = "EuPathDB",
  overwrite = FALSE)
}
\arguments{
\item{gids}{Gene IDs to cross reference against uniprot.}

\item{entry}{Eupath entry to cross reference.}

\item{dir}{Working directory.}

\item{overwrite}{Overwrite the savefile.}
}
\description{
I have been working on getting stupid v42 eupathdb into annotationhub
forever, it is not particularly interesting to me, and for reasons passing
all understanding, it seems impossible to get it to work properly for Lori.
In the mean time, I had some ideas of fun things to do.
}
\details{
Notably, my other package 'hpgltools' has a nifty uniprot downloader/parser.
It may reasonably easily be used to bring into an orgdb the set of uniprot
annotations.

This is just a silly initial implementation because I have been sipping on
whisky.  But I think it gets the idea across.
}
