% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_metadata.R
\name{get_all_metadata}
\alias{get_all_metadata}
\title{Invoke download_eupath_metadata() using all the sub-projects of the EuPathDB.}
\usage{
get_all_metadata(webservice = "all")
}
\arguments{
\item{webservice}{Assume all services are desired.}
}
\description{
Invoke download_eupath_metadata() using all the sub-projects of the EuPathDB.
}
