% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rda.R
\name{get_rda_filename}
\alias{get_rda_filename}
\title{Provide the filename for a temporary rda file for saving data while running.}
\usage{
get_rda_filename(entry, name)
}
\arguments{
\item{entry}{metadata entry to define a tablename.}

\item{name}{Specific table to provide.}
}
\description{
This defines the names of the various savefiles when saving tables.
}
