% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eupath_orthologs.R
\name{load_eupath_pkg}
\alias{load_eupath_pkg}
\title{Loads a pkg into the current R environment.}
\usage{
load_eupath_pkg(name, webservice = "eupathdb")
}
\description{
Loads a pkg into the current R environment.
}
