% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_orgdb_annotations.R
\name{load_orgdb_annotations}
\alias{load_orgdb_annotations}
\title{Load organism annotation data from an orgdb sqlite package.}
\usage{
load_orgdb_annotations(
  orgdb = NULL,
  gene_ids = NULL,
  include_go = FALSE,
  keytype = "gid",
  location_column = "annot_location_text",
  type_column = "annot_gene_type",
  name_column = "annot_gene_product",
  fields = NULL,
  sum_exon_widths = FALSE
)
}
\arguments{
\item{orgdb}{OrganismDb instance.}

\item{gene_ids}{Search for a specific set of genes?}

\item{include_go}{Ask the Dbi for gene ontology information?}

\item{keytype}{Primary key of the tables, 'gid' for EuPathDB data.}

\item{location_column}{Which column contains the location data for the genes?}

\item{type_column}{Use this column to identify the gene type.}

\item{name_column}{Use this column to identify the gene name.}

\item{fields}{Columns included in the output.}

\item{sum_exon_widths}{Perform a sum of the exons in the data set?}
}
\value{
Table of geneids, chromosomes, descriptions, strands, types, and lengths.
}
\description{
Creates a dataframe gene and transcript information for a given set of gene
ids using the AnnotationDbi interface.
}
\details{
This defaults to a few fields which I have found most useful, but the brave
or pathological can pass it 'all'.
}
\examples{
\dontrun{
 one_gene <- load_orgdb_annotations(org, c("LmJF.01.0010"))
}
}
\seealso{
\pkg{AnnotationDbi} \pkg{GenomicFeatures} \pkg{BiocGenerics}
 \code{\link[AnnotationDbi]{columns}} \code{\link[AnnotationDbi]{keytypes}}
 \code{\link[AnnotationDbi]{select}} \code{\link[GenomicFeatures]{exonsBy}}
}
\author{
atb
}
