% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_bsgenome.R
\name{make_eupath_bsgenome}
\alias{make_eupath_bsgenome}
\title{Generate a BSgenome package from the eupathdb.}
\usage{
make_eupath_bsgenome(entry, version = NULL, dir = "EuPathDB",
  copy_s3 = FALSE, installp = TRUE, reinstall = FALSE, ...)
}
\arguments{
\item{entry}{Single eupathdb metadata entry.}

\item{version}{Which version of the eupathdb to use for creating the BSGenome?}

\item{dir}{Working directory.}

\item{copy_s3}{Copy the 2bit file into an s3 staging directory for copying to AnnotationHub?}

\item{reinstall}{Rewrite an existing package directory.}

\item{...}{Extra arguments for downloading metadata when not provided.}
}
\value{
List of package names generated (only 1).
}
\description{
Since we go to the trouble to try and generate nice orgdb/txdb/organismdbi
packages, it seems to me that we ought to also be able to make a readable
genome package.  I should probably use some of the logic from this to make
the organismdbi generator smarter.
}
\author{
atb
}
