% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{make_eupath_txdb}
\alias{make_eupath_txdb}
\title{Generate TxDb for EuPathDB organism}
\usage{
make_eupath_txdb(entry = NULL, dir = "eupathdb", version = NULL,
  reinstall = FALSE, ...)
}
\arguments{
\item{entry}{One row from the organism metadata.}

\item{dir}{Base directory for building the package.}

\item{version}{Which version of the eupathdb to use for creating this package?}

\item{reinstall}{Overwrite an existing installed package?}

\item{...}{Extra arguments for getting metadata.}

\item{species}{Species name from the eupathdb metadata.}
}
\value{
TxDb instance name.
}
\description{
Generate TxDb for EuPathDB organism
}
\author{
Keith Hughitt with significant modifications by atb.
}
