% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_txdb.R
\name{make_eupath_txdb}
\alias{make_eupath_txdb}
\title{Generate an EuPathDB organism TxDb package.}
\usage{
make_eupath_txdb(entry = NULL, dir = "EuPathDB", version = NULL,
  reinstall = FALSE)
}
\arguments{
\item{entry}{One row from the organism metadata.}

\item{dir}{Base directory for building the package.}

\item{version}{Which version of the eupathdb to use for creating this package?}

\item{reinstall}{Overwrite an existing installed package?}
}
\value{
TxDb instance name.
}
\description{
This will hopefully create a txdb package and granges savefile for a single
species in the eupathdb.  This depends pretty much entirely on the successful
download of a gff file from the eupathdb.
}
\author{
Keith Hughitt with significant modifications by atb.
}
