% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_final_package.R
\name{move_final_package}
\alias{move_final_package}
\title{Move a package file to its final location for collation by AnnotationHubData.}
\usage{
move_final_package(pkgname, type = "orgdb", dir = "EuPathDB")
}
\description{
Move a package file to its final location for collation by AnnotationHubData.
}
