% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_annotations.R
\name{post_eupath_annotations}
\alias{post_eupath_annotations}
\title{Create a series of POST requests which download all the annotation data for a species.}
\usage{
post_eupath_annotations(
  entry = NULL,
  overwrite = FALSE,
  build_dir = "EuPathDB"
)
}
\arguments{
\item{entry}{Eupathdb annotation entry.}

\item{overwrite}{Overwrite existing data if it exists?}

\item{build_dir}{Location to dump the resulting data.}
}
\description{
The only way I have figured out how to download mass data from the eupathdb
is to ask for a raw dump of all available data using the GenesByGeneType
WADL.  Therefore, this function iterates over the various sequence types that
I have noticed at the eupathdb and does that for each type.
}
