% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_go.R
\name{post_eupath_go_table}
\alias{post_eupath_go_table}
\title{Use the POST interface to get GO data from the EuPathDB.}
\usage{
post_eupath_go_table(
  entry = NULL,
  build_dir = "EuPathDB",
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{build_dir}{Location to write savefiles.}

\item{overwrite}{Overwrite intermediate savefiles in case of incomplete install?}
}
\value{
A big honking table.
}
\description{
Use the POST interface to get GO data from the EuPathDB.
}
