% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_eupath_pathway.R
\name{post_eupath_pathway_table}
\alias{post_eupath_pathway_table}
\title{Use the post interface to get pathway data.}
\usage{
post_eupath_pathway_table(
  entry,
  working_species,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{entry}{The full annotation entry.}

\item{overwrite}{If trying again, overwrite the savefile?}

\item{build_dir}{Location to which to save intermediate savefile.}
}
\value{
A big honking table.
}
\description{
Use the post interface to get pathway data.
}
