% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_s3_file.R
\name{query_s3_file}
\alias{query_s3_file}
\title{Perform what should be a completely silly final check on the file which is to be copied to s3.}
\usage{
query_s3_file(row, file_type = "OrgDb", file_column = "OrgdbFile")
}
\arguments{
\item{row}{Metadata row to query.}

\item{file_type}{Currently I have 3 file types of interest.}

\item{file_column}{Name of the column with the locations of the files of
interest.}
}
\value{
MD5 checksum of the resulting file, or NULL.
}
\description{
This function really should not be needed.  But damn.  This will do a final check that the
data in the s3 staging directory is loadable in R and return the md5 sum of the file.
Thus the md5 sum may be added to the metadata.
}
