% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_s3_file.R
\name{query_s3_file}
\alias{query_s3_file}
\title{Perform what should be a completely silly final check on the file which is to be copied to s3.
This function really should not be needed.  But damn.  This will do a final check that the
data in the s3 staging directory is loadable in R and return the md5 sum of the file.
Thus the md5 sum may be added to the metadata.}
\usage{
query_s3_file(row, file_type = "OrgDb", file_column = "OrgdbFile")
}
\arguments{
\item{row}{Metadata row to query.}

\item{file_type}{Currently I have 3 file types of interest.}

\item{file_column}{Column of the metadata to use to get the filename.}
}
\description{
Perform what should be a completely silly final check on the file which is to be copied to s3.
This function really should not be needed.  But damn.  This will do a final check that the
data in the s3 staging directory is loadable in R and return the md5 sum of the file.
Thus the md5 sum may be added to the metadata.
}
