% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eupath_orgdb.R, R/remove_eupath_nas.R
\name{remove_eupath_nas}
\alias{remove_eupath_nas}
\title{Get rid of spurious NA entries in a table from the eupathdb.}
\usage{
remove_eupath_nas(table, name = "annot")

remove_eupath_nas(table, name = "annot")
}
\arguments{
\item{table}{dataframe of SQLite-bound data}

\item{name}{column prefix, just used for printing for the moment.}
}
\description{
If these are not removed, creating the sqlite will fail.

If these are not removed, creating the sqlite will fail.
}
