% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xref_taxonomy.R
\name{search_na_taxon}
\alias{search_na_taxon}
\title{Try harder to fill in the taxonomy ID number.}
\usage{
search_na_taxon(
  metadatum,
  all_taxa_ids,
  metadata_taxon_column = "TaxonUnmodified",
  taxon_number_column = "TaxonomyID"
)
}
\description{
Try harder to fill in the taxonomy ID number.
}
