% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupathdb.R
\name{start_eupathdb}
\alias{start_eupathdb}
\alias{availableEuPathDB}
\title{Get started with EuPathDB}
\usage{
start_eupathdb(type = "GRanges")
}
\arguments{
\item{type}{Choose this type of metadatum to open.}
}
\value{
Used for its side-effect of opening the package vignette. A
        vector of experiment identifiers.
}
\description{
This function has always been here.  To be honest, I am not completely sure of its purpose.
}
\examples{
start_eupathdb()
}
\author{
Keith Hughitt
}
