% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eupath_metadata.R
\name{write_eupath_metadata}
\alias{write_eupath_metadata}
\title{Standardize the writing of csv metadata.}
\usage{
write_eupath_metadata(valid_metadata, service, bioc_version, eu_version)
}
\arguments{
\item{valid_metadata}{Set of metadata which are deemed 'valid.'}

\item{service}{EupathDB subproject, or the set of all projects named
'eupathdb'.}

\item{bioc_version}{Version of Bioconductor used for this set of metadata.}

\item{eu_version}{Version of the EuPathDB used for this set of metadata.}
}
\value{
List containing the filenames written.
}
\description{
Here is an attempt at making the various written csv files for AnnotationHub
consistent.  I had a problem when filtering the data in which I managed to
reverse the version numbers for bioconductor and eupathdb.  I am very smrt.
}
