% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_eupath_metadata.R
\name{write_eupath_metadata}
\alias{write_eupath_metadata}
\title{Standardize the writing of csv metadata.}
\usage{
write_eupath_metadata(
  metadata,
  service = "eupathdb",
  type = "valid",
  bioc_version = "3.9",
  eu_version = "44"
)
}
\arguments{
\item{metadata}{Set of metadata.}

\item{service}{EupathDB subproject, or the set of all projects named
'eupathdb'.}

\item{type}{Either valid or invalid, defines the final output filenames.}

\item{bioc_version}{Version of Bioconductor used for this set of metadata.}

\item{eu_version}{Version of the EuPathDB used for this set of metadata.}
}
\value{
List containing the filenames written.
}
\description{
This function effectively splits the metadata from a single data frame to a set of
individual files, one for each data type created.
}
